DimosoProject/DimosoProject/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
Babel==2.9.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.2.5
django-ckeditor==6.2.0
django-crispy-forms==1.12.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==1.2.2
docutils==0.17.1
future==0.18.2
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.2
imagesize==1.2.0
incremental==21.3.0
Jinja2==3.0.2
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.0
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
Twisted==22.1.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.0.1
urllib3==1.26.6
webencodings==0.5.1
zope.interface==5.4.0
